Refs #32508 -- Raised Type/ValueError instead of using "assert" in django.db.models.

+++ django/db/models/base.py
@@ -947,12 +947,12 @@ class Model(metaclass=ModelBase):
                     field.delete_cached_value(self)
 
     def delete(self, using=None, keep_parents=False):
+        if self.pk is None:
+            raise ValueError(
+                "%s object can't be deleted because its %s attribute is set "
+                "to None." % (self._meta.object_name, self._meta.pk.attname)
+            )
         using = using or router.db_for_write(self.__class__, instance=self)
-        assert self.pk is not None, (
-            "%s object can't be deleted because its %s attribute is set to None." %
-            (self._meta.object_name, self._meta.pk.attname)
-        )
-
         collector = Collector(using=using)
         collector.collect([self], keep_parents=keep_parents)
         return collector.delete()
